interpolateCR change float to double

+++ osg4web/bzr_import/include/CommonCore/SceneHandlers/AnimateViewHandler.h
@@ -51,7 +51,7 @@ namespace SceneHandlers
 		bool handle(const osgGA::GUIEventAdapter& ea,osgGA::GUIActionAdapter& aa);
 
 		// Catmull-Rom Interpolation (p1,p2)
-		osg::Vec3 interpolateCR(float t, osg::Vec3 p0, osg::Vec3 p1, osg::Vec3 p2, osg::Vec3 p3);
+		osg::Vec3 interpolateCR(double t, osg::Vec3 p0, osg::Vec3 p1, osg::Vec3 p2, osg::Vec3 p3);
 
 		//Ridefinizioni del Gestore dei Comandi
 		virtual std::string handleAction(std::string argument);
